runinhost_once.sh only run once

+++ mysql5/runinhost_once.sh
@@ -2,17 +2,21 @@
 
 tdir=/opt/dockerdata/mysql5
 
-mkdir -p "$tdir/log"
+if [ ! -d $tdir ]; then
 
-cp supervisord.conf  $tdir
-cp my.cnf $tdir
+  mkdir -p "$tdir/log"
 
-cp addn-hosts $tdir
-cp dnsmasq.conf $tdir
+  cp supervisord.conf  $tdir
+  cp my.cnf $tdir
 
-if [ ! -d "$tdir/mysql" ]; then
-  cp mysqllib.5173.tar.gz $tdir
-  cd $tdir
-  tar -zxf mysqllib.5173.tar.gz
-  rm -rf mysqllib.5173.tar.gz
+  cp addn-hosts $tdir
+  cp dnsmasq.conf $tdir
+
+  if [ ! -d "$tdir/mysql" ]; then
+    cp mysqllib.5173.tar.gz $tdir
+    cd $tdir
+    tar -zxf mysqllib.5173.tar.gz
+    rm -rf mysqllib.5173.tar.gz
+  fi
+
 fi
